fix(AYC-000): auto-save unsaved editor changes before artifact export
